bugfixes and moved base code for post and patch transactions to saveOrUpdateTransactions service method. Moved SuccessErrorMessage and DataError objects into more appropriately named HTTP response objects, migrated Vendor and paymethod objects to hold foreign keys to each other's objects, need to cleanup the database next

+++ src/main/java/com/jack/utility/General.java
@@ -10,8 +10,16 @@ public class General {
     }
 
     /* Algorithm Utility */
-    public static int mergeHash(int hash1, int hash2) {
-        String toHash = String.format("%032d", hash1)+ String.format("%032d", hash2);
+
+    /**
+     * Creates a String by merging two longs at constant length of 19
+     * and then hashes the result into a unique hash
+     * @param hash1
+     * @param hash2
+     * @return
+     */
+    public static int mergeHash(long hash1, long hash2) {
+        String toHash = String.format("%019d", hash1)+ String.format("%019d", hash2);
          return Math.abs(toHash.hashCode());
     }
